+++ app/Http/Controllers/SaleManController.php
@@ -164,6 +164,10 @@ class SaleManController extends Controller
             toast('غير مسموح بحذف بياناتك ','danger');
             return back();
         }
+        if (Auth::user()->bills->count() > 0){
+            toast('غير مسموح بحذف بياناتك ','danger');
+            return back();
+        }
         $user= User::findOrFail($id);
         $user->delete();
         toast('تم الحذف بنجاح','success');
